requirement.txt
pymongo
python-dotenv
motor
fastapi
uvicorn
pandas
numpy
scikit-learn
matplotlib
